+++ test/lightning_web/live/workflow_live/edit_test.exs
@@ -10,9 +10,11 @@ defmodule LightningWeb.WorkflowLive.EditTest do
 
   alias Lightning.Helpers
   alias Lightning.Repo
+  # alias Lightning.Run
   alias Lightning.Workflows
   alias Lightning.Workflows.Snapshot
   alias Lightning.Workflows.Workflow
+  # alias Lightning.WorkOrder
   alias LightningWeb.CredentialLiveHelpers
 
   setup :register_and_log_in_user
@@ -1985,5 +1987,167 @@ defmodule LightningWeb.WorkflowLive.EditTest do
 
       assert has_element?(view, "#ai-assistant-error", error_message)
     end
+  end
+
+  describe "Allow low priority access users to retry steps and create workorders" do
+    setup do
+      project = insert(:project)
+
+      high_priority_user =
+        insert(:user,
+          email: "amy@openfn.org",
+          first_name: "Amy",
+          last_name: "Ly"
+        )
+
+      low_priority_user =
+        insert(:user,
+          email: "ana@openfn.org",
+          first_name: "Ana",
+          last_name: "Ba"
+        )
+
+      insert(:project_user,
+        project: project,
+        user: high_priority_user,
+        role: :admin
+      )
+
+      insert(:project_user,
+        project: project,
+        user: low_priority_user,
+        role: :admin
+      )
+
+      workflow = insert(:simple_workflow, project: project)
+
+      {:ok, snapshot} = Snapshot.get_or_create_latest_for(workflow)
+
+      %{jobs: [job], triggers: [trigger]} = workflow
+
+      [input_dataclip, output_dataclip] =
+        insert_list(2, :dataclip,
+          body: %{player: "sadio mane"},
+          project: workflow.project
+        )
+
+      %{runs: [run]} =
+        insert(:workorder,
+          trigger: trigger,
+          dataclip: input_dataclip,
+          workflow: workflow,
+          snapshot: snapshot,
+          state: :success,
+          runs: [
+            build(:run,
+              starting_trigger: trigger,
+              dataclip: input_dataclip,
+              steps: [
+                build(:step,
+                  input_dataclip: input_dataclip,
+                  output_dataclip: output_dataclip,
+                  job: job,
+                  inserted_at: Timex.now() |> Timex.shift(seconds: -10),
+                  started_at: Timex.now() |> Timex.shift(seconds: -10),
+                  snapshot: snapshot
+                )
+              ],
+              inserted_at: Timex.now() |> Timex.shift(seconds: -12),
+              snapshot: snapshot,
+              state: :success
+            )
+          ]
+        )
+
+      %{
+        project: project,
+        high_priority_user: high_priority_user,
+        low_priority_user: low_priority_user,
+        workflow: workflow,
+        snapshot: snapshot,
+        run: run,
+        job: job
+      }
+    end
+
+    test "Users with low priority access to the workflow canvas will automatically be locked in a snapshot when the high prior uses saves the workflow",
+         %{
+           conn: conn,
+           project: project,
+           workflow: workflow,
+           snapshot: snapshot,
+           run: run,
+           job: job,
+           high_priority_user: high_priority_user,
+           low_priority_user: low_priority_user
+         } do
+      {high_priority_view, low_priority_view} =
+        access_views(
+          conn,
+          project,
+          workflow,
+          run,
+          job,
+          high_priority_user,
+          low_priority_user
+        )
+
+      assert high_priority_view
+             |> has_element?("#inspector-workflow-version", "latest")
+
+      assert low_priority_view
+             |> has_element?("#inspector-workflow-version", "latest")
+
+      high_priority_view |> select_node(%{id: job.id})
+
+      high_priority_view |> click_edit(%{id: job.id})
+
+      high_priority_view |> change_editor_text("Job expression 1")
+
+      high_priority_view |> form("#workflow-form") |> render_submit()
+
+      assert high_priority_view
+             |> has_element?("#inspector-workflow-version", "latest")
+
+      refute low_priority_view
+             |> has_element?("#inspector-workflow-version", "latest")
+
+      assert low_priority_view
+             |> has_element?(
+               "#inspector-workflow-version",
+               "#{String.slice(snapshot.id, 0..6)}"
+             )
+
+      assert low_priority_view |> render() =~
+               "This workflow has been updated. You&#39;re no longer on the latest version."
+
+      workflow = Repo.reload(workflow)
+
+      assert workflow.lock_version == snapshot.lock_version + 1
+    end
+  end
+
+  defp access_views(
+         conn,
+         project,
+         workflow,
+         run,
+         job,
+         high_priority_user,
+         low_priority_user
+       ) do
+    {:ok, high_priority_view, _html} =
+      live(
+        log_in_user(conn, high_priority_user),
+        ~p"/projects/#{project}/w/#{workflow}?#{[a: run, s: job, m: "expand"]}"
+      )
+
+    {:ok, low_priority_view, _html} =
+      live(
+        log_in_user(conn, low_priority_user),
+        ~p"/projects/#{project}/w/#{workflow}?#{[a: run, s: job, m: "expand"]}"
+      )
+
+    {high_priority_view, low_priority_view}
   end
 end